fix telegram methods

+++ src/telegram/commands/meme.ts
@@ -1,5 +1,6 @@
-import { apiLeagueInstance } from '../../http/apiLeague.ts';
 import type { CommandHandlerArgs } from '../../types.ts';
+import { getApiLeagueMeme } from '../../utils/getApiLeagueMeme.ts';
+import { logger } from '../../utils/logger.ts';
 
 export async function memeCommand({ api, msg }: CommandHandlerArgs) {
   const chatId = msg.chat.id;
@@ -11,41 +12,34 @@ export async function memeCommand({ api, msg }: CommandHandlerArgs) {
       delay: 0
     });
 
-    const sendMessageResponse = await api.sendTelegramMessage({
+    const searchMessageResponse = await api.sendTelegramMessage({
       chatId,
       text: `Ищу мем 🤔`
     });
 
-    const apiLeagueResponse = await apiLeagueInstance.get(
-      '/retrieve-random-meme',
-      {
-        params: {
-          minRating: 1,
-          maxAgeDays: 5
-        }
-      }
-    );
+    const { memeUrl, text } = await getApiLeagueMeme();
 
-    const requestCountLeft = apiLeagueResponse.headers['x-api-quota-left'];
-    const requestCountUsed = apiLeagueResponse.headers['x-api-quota-used'];
-
-    const memeDescription = apiLeagueResponse.data.description;
-    const memUrl = apiLeagueResponse.data.url;
-
-    const text = `🐸 Описание мема: ${memeDescription}\n\nОсталось мемов на сегодня ${requestCountLeft}\nИспользовано мемов ${requestCountUsed}`;
+    if (!memeUrl) {
+      return await api.sendTelegramMessage({
+        chatId,
+        text: 'Не удалось найти мем 🤡'
+      });
+    }
 
     await api.sendPhoto({
       chatId,
-      img: memUrl,
+      img: memeUrl,
       text: text
     });
 
     await api.deleteTelegramMessageWithDelay({
-      messageId: sendMessageResponse.message_id,
+      messageId: searchMessageResponse.message_id,
       chatId,
       delay: 0
     });
-  } catch {
+  } catch (e) {
+    logger('error', e);
+
     await api.sendTelegramMessage({
       chatId,
       text: 'Не удалось отправить мем 🤡'
